requirements.txt
ephem==4.1.4
pyaes==1.6.1
pyasn1==0.4.8
pylunar==0.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
rsa==4.9
six==1.16.0
Telethon==1.27.0
